binance-prediction/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.25.0
python-binance>=1.0.15
flask>=2.0.0
jupyter>=1.0.0
joblib>=1.0.0
python-dotenv>=0.19.0
